Move `getPlaybackRatesArray` to utils to avoid audio-player importing video-player
